+++ cogito/put_test.go
@@ -7,7 +7,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/hashicorp/go-hclog"
 	"gotest.tools/v3/assert"
 
 	"github.com/Pix4D/cogito/cogito"
@@ -65,7 +64,7 @@ func (ms MockSinker) Send() error {
 func TestPutSuccess(t *testing.T) {
 	putter := MockPutter{sinkers: []cogito.Sinker{MockSinker{}}}
 
-	err := cogito.Put(hclog.NewNullLogger(), nil, nil, nil, putter)
+	err := cogito.Put(testhelp.MakeTestLog(), nil, nil, nil, putter)
 
 	assert.NilError(t, err)
 }
@@ -78,7 +77,7 @@ func TestPutFailure(t *testing.T) {
 	}
 
 	test := func(t *testing.T, tc testCase) {
-		err := cogito.Put(hclog.NewNullLogger(), nil, nil, nil, tc.putter)
+		err := cogito.Put(testhelp.MakeTestLog(), nil, nil, nil, tc.putter)
 
 		assert.ErrorContains(t, err, tc.wantErr)
 	}
@@ -124,7 +123,7 @@ func TestPutFailure(t *testing.T) {
 
 func TestPutterLoadConfigurationSuccess(t *testing.T) {
 	in := testhelp.ToJSON(t, basePutRequest)
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 	err := putter.LoadConfiguration(in, []string{"dummy-dir"})
 
@@ -143,7 +142,7 @@ func TestPutterLoadConfigurationSinksOverrideSuccess(t *testing.T) {
 	},
 	  "params": {"sinks": ["gchat"]}
 	}`)
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 	inputDir := []string{""}
 	err := putter.LoadConfiguration(in, inputDir)
 	assert.NilError(t, err)
@@ -164,7 +163,7 @@ func TestPutterLoadConfigurationFailure(t *testing.T) {
 
 	test := func(t *testing.T, tc testCase) {
 		in := testhelp.ToJSON(t, tc.putInput)
-		putter := cogito.NewPutter(hclog.NewNullLogger())
+		putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 		err := putter.LoadConfiguration(in, tc.args)
 
@@ -205,7 +204,7 @@ func TestPutterLoadConfigurationInvalidParamsFailure(t *testing.T) {
   "params": {"pizza": "margherita"}
 }`)
 	wantErr := `put: parsing request: json: unknown field "pizza"`
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 	err := putter.LoadConfiguration(in, nil)
 
@@ -219,7 +218,7 @@ func TestPutterLoadConfigurationMissingGchatwebHook(t *testing.T) {
   "params": {}
 }`)
 	wantErr := `put: source: missing keys: gchat_webhook`
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 	err := putter.LoadConfiguration(in, nil)
 
@@ -233,7 +232,7 @@ func TestPutterLoadConfigurationUnknownSink(t *testing.T) {
   "params": {}
 }`)
 	wantErr := `put: source: invalid sink(s): [pizza]`
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 	err := putter.LoadConfiguration(in, nil)
 
@@ -247,7 +246,7 @@ func TestPutterLoadConfigurationUnknownSinkPutParams(t *testing.T) {
   "params": {"sinks": ["pizza"]}
 }`)
 	wantErr := `put: arguments: unsupported sink(s): [pizza]`
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 	err := putter.LoadConfiguration(in, nil)
 
@@ -263,7 +262,7 @@ func TestPutterProcessInputDirSuccess(t *testing.T) {
 	}
 
 	test := func(t *testing.T, tc testCase) {
-		putter := cogito.NewPutter(hclog.NewNullLogger())
+		putter := cogito.NewPutter(testhelp.MakeTestLog())
 		tmpDir := testhelp.MakeGitRepoFromTestdata(t, tc.inputDir,
 			"https://github.com/dummy-owner/dummy-repo", "dummySHA", "banana")
 		putter.InputDir = filepath.Join(tmpDir, filepath.Base(tc.inputDir))
@@ -318,7 +317,7 @@ func TestPutterProcessInputDirFailure(t *testing.T) {
 	test := func(t *testing.T, tc testCase) {
 		tmpDir := testhelp.MakeGitRepoFromTestdata(t, tc.inputDir,
 			"https://github.com/dummy-owner/dummy-repo", "dummySHA", "banana mango")
-		putter := cogito.NewPutter(hclog.NewNullLogger())
+		putter := cogito.NewPutter(testhelp.MakeTestLog())
 		putter.Request = cogito.PutRequest{
 			Source: cogito.Source{GhHostname: github.GhDefaultHostname, Owner: "dummy-owner", Repo: "dummy-repo"},
 			Params: tc.params,
@@ -395,7 +394,7 @@ func TestPutterProcessInputDirNonExisting(t *testing.T) {
 }
 
 func TestPutterSinks(t *testing.T) {
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 	sinks := putter.Sinks()
 	assert.Assert(t, len(sinks) == 2)
@@ -407,7 +406,7 @@ func TestPutterSinks(t *testing.T) {
 }
 
 func TestPutterCustomSinks(t *testing.T) {
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 	putter.Request = cogito.PutRequest{
 		Params: cogito.PutParams{Sinks: []string{"gchat"}},
 	}
@@ -418,7 +417,7 @@ func TestPutterCustomSinks(t *testing.T) {
 }
 
 func TestPutterOutputSuccess(t *testing.T) {
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 	err := putter.Output(io.Discard)
 
@@ -426,7 +425,7 @@ func TestPutterOutputSuccess(t *testing.T) {
 }
 
 func TestPutterOutputFailure(t *testing.T) {
-	putter := cogito.NewPutter(hclog.NewNullLogger())
+	putter := cogito.NewPutter(testhelp.MakeTestLog())
 
 	err := putter.Output(&testhelp.FailingWriter{})